requirements.txt
alembic==1.13.1
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email-validator==1.1.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.4.4
Flask-Validator==1.4.2
greenlet==3.0.3
idna==3.7
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
py-moneyed==0.8.0
pycountry==23.12.11
python-dotenv==1.0.1
pytz==2020.1
schwifty==2020.9.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
